🔥 delete functions file
