+++ lib/functions.ts
@@ -1,44 +1,49 @@
-import type {Experience, MarkdownData, Project, Skill} from "./types";
-import { readdirSync, readFileSync } from "fs";
-import { join } from "path";
-import { serialize } from "next-mdx-remote/serialize";
-import matter from "gray-matter";
-
-export async function readMarkdownFiles<
-  T extends {
-    [key: string]: any;
-  }
->(...paths: string[]): Promise<MarkdownData<T>[]> {
-  const dataList: MarkdownData<T>[] = [];
-
-  const markdownContent = readdirSync(join(process.cwd(), ...paths)).map(
-    (file) => {
-      return readFileSync(join(process.cwd(), ...paths, file), "utf8");
+/**
+ * Petit utilitaire pour chainer les classes css en react tout en évitant
+ *  les "false" et "null" dans les classes.
+ *
+ *  @example
+ *    const classes = clsx(
+ *        "class1",
+ *        "class2",
+ *        undefined,
+ *        {
+ *          class3: true,
+ *          class4: false,
+ *        });
+ *     // retournera "class1 class2 class3"
+ *
+ * @param args
+ */
+export function clsx(
+  ...args: (string | undefined | Record<string, boolean>)[]
+): string {
+  const classes: string[] = [];
+  for (const arg of args) {
+    switch (typeof arg) {
+      case "string":
+        if (arg) {
+          classes.push(arg);
+        }
+        break;
+      case "object":
+        if (arg) {
+          for (const key in arg) {
+            if (arg[key]) {
+              classes.push(key);
+            }
+          }
+        }
+        break;
     }
-  );
-
-  for (const fileContent of markdownContent) {
-    const { content, data } = matter(fileContent);
-    const serializedContent = await serialize(content);
-
-    dataList.push({
-      data: data as T,
-      content,
-      serializedContent,
-    });
   }
-
-  return dataList;
+  return classes.join(" ");
 }
 
-export async function getExperiences(): Promise<MarkdownData<Experience>[]> {
-  return readMarkdownFiles<Experience>("data", "experience");
-}
-
-export function getProjects(): Promise<MarkdownData<Project>[]> {
-  return readMarkdownFiles<Project>("data", "projects");
-}
-
-export function getSkills(): Promise<MarkdownData<Skill>[]> {
-  return readMarkdownFiles<Skill>("data", "skills");
+export function formatDate(dateStr: string | Date): string {
+  return new Intl.DateTimeFormat("fr-FR", {
+    month: "short",
+    day: "2-digit",
+    year: "numeric",
+  }).format(new Date(dateStr));
 }
