--- lib/webutils.ts
@@ -1,49 +1,0 @@
-/**
- * Petit utilitaire pour chainer les classes css en react tout en évitant
- *  les "false" et "null" dans les classes.
- *
- *  @example
- *    const classes = clsx(
- *        "class1",
- *        "class2",
- *        undefined,
- *        {
- *          class3: true,
- *          class4: false,
- *        });
- *     // retournera "class1 class2 class3"
- *
- * @param args
- */
-export function clsx(
-  ...args: (string | undefined | Record<string, boolean>)[]
-): string {
-  const classes: string[] = [];
-  for (const arg of args) {
-    switch (typeof arg) {
-      case "string":
-        if (arg) {
-          classes.push(arg);
-        }
-        break;
-      case "object":
-        if (arg) {
-          for (const key in arg) {
-            if (arg[key]) {
-              classes.push(key);
-            }
-          }
-        }
-        break;
-    }
-  }
-  return classes.join(" ");
-}
-
-export function formatDate(dateStr: string | Date): string {
-  return new Intl.DateTimeFormat("fr-FR", {
-    month: "short",
-    day: "2-digit",
-    year: "numeric",
-  }).format(new Date(dateStr));
-}
